refactor(mail): flatten jmap and stalwart packages to single modules
